# Don't use an adjacency matrix for choose.edges. Way, way faster.