This makes sim.rand.graph.clust significantly faster and much less memory hungry.

The adjacency matrix was only really used for pretty standard functions that igraph supplies anyway.

For large graphs, this makes it way faster. Comparison on a graph with 10000 vertices:

```
  expression   min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory time 
  <bch:expr> <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list> <lis>
1 test         27s    27s    0.0370    25.4GB    0.407     1    11        27s
1 test       290ms  311ms      3.22    36.1MB     4.83     2     3      621ms
```

So about 100 times faster, and around 1000 times less memory allocation. Rejoice! :)