@cwatson 2021-03-26
Thanks for the contribution! I had never tested on a graph that large before, and these functions were written pretty early in the package's history.

Could you provide some more benchmark info? Specifically:

- what version of `R` and `igraph` are you using? I seem to remember back when I worked on these functions, using the adjacency matrix was much faster but `igraph` has gone through a lot of changes since then.
- have you tested on graphs with variable degree and clustering? I'd be interested to see if you see the same kind of improvement with dense and/or highly-clustered graphs.

It'll take me some time to merge this; I have a few bug fixes that I would like to get to first.