@cwatson 2021-03-27
I did indeed use `igraph` native functions when I first wrote those functions; I changed them to use the adjacency matrix about 4 years ago. I recall it having been much *slower* with the graph functions, so if you don't mind testing smaller graphs (I think 10K is larger than almost any neuroimaging application) and vary the density and clustering, that would be really helpful.

If it is only beneficial for very large graphs, it might make sense to have different functionality based on the vertex count.